requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
sqlalchemy>=2.0.0
asyncpg
alembic
python-dotenv
redis>=4.5.0
aioredis>=2.0.0 